Create README.md
This script contains a conditional structure that maps a student's numerical grade to a letter grade (A–D) based on semester performance in a subject. As students taking tests during the semester, there are 4 major tests that will determine what letter grade they will receive at the end of semester. This script will do the math by inputting numeric grades from the 4 tests, and the program will do the calculation and will give a letter grade as output